Implement adding a mask to the databse

+++ src/inventoryapp/forms.py
@@ -1,10 +1,10 @@
-from wtforms import SelectField, StringField, DecimalField, BooleanField, validators
+from wtforms import SelectField, StringField, DecimalField, IntegerField, BooleanField, validators
 from flask_wtf import Form
 
 class NewMaskForm(Form):
     mask_name = StringField('Mask Name', validators=[validators.Length(min=4, max=25)])
     form_factor = StringField('Form Factor', validators=[validators.Length(min=3, max=25)])
-    quantity = DecimalField('Initial Quantity', validators=[validators.NumberRange(min=0)])
+    quantity = IntegerField('Initial Quantity', validators=[validators.NumberRange(min=0)])
 
 
 class EditMaskForm(Form):
@@ -13,11 +13,11 @@ class EditMaskForm(Form):
 
 
 class AddCardsForm(Form):
-    quantity = DecimalField('Quantity', validators=[validators.NumberRange(min=1)])
+    quantity = IntegerField('Quantity', validators=[validators.NumberRange(min=1)])
 
 
 class TransferCardsForm(Form):
-    quantity = DecimalField('Quantity', validators=[validators.NumberRange(min=1)])
+    quantity = IntegerField('Quantity', validators=[validators.NumberRange(min=1)])
     destination = SelectField('Destination', coerce=int)
 
 class EmptyTrashForm(Form):
